split models file, still have circular import
