+++ .gitignore
@@ -155,4 +155,4 @@ docs/
 test-napari.pub
 data/
 BentoML/
-models/
+
