+++ src/server/dcp_server/config.cfg
@@ -15,6 +15,7 @@
 
     "model": { 
         "segmentor": {
+            "model_class": "CustomCellposeModel",
             "model_type": "cyto"
             
         },
